fix: set dev docker-compose network as external
This network is used across multiple apps/services by Data
Infrastructure, but if the network is not declared as external then it
is removed when `docker-compose down` is run - which will automatically
bring down any other compose projects that are still running on the same
network. This isn't desirable. If we declare the network as external,
compose will expect it to exist and not remove it when bringing down the
project, which is more desirable. The only caveat is that devs will need
to manually created the network before running the project - but if they
haven't done this, docker-compose throws an error and provides the
solution, so this is not a barrier.