build(deps): move to requirements.txt from Conda
While it all worked in Docker, it didn't work in PaaS - gunicorn seemed to get
stuck in an infinite "Booting worker" loop.

After SSH-ing into the instance and trying to find what was causing this, it
seemed to be a segmentation fault, and a segmentation caused by importing
sqlalchemy. No real idea what caused it, but wondering if it's do with with the
multiple Python versions installed when using Conda: PaaS's, and Conda's.

Most of our apps use the requirements.in/txt pattern, so moving to that pattern
- we're more familiar with it and there is an argument it's "simpler"